requirements.txt
api==0.0.7
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2.10
django-allauth==0.50.0
django-bootstrap-v5==1.0.11
django-cors-headers==3.12.0
django-extensions==3.1.5
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.3
mypy-extensions==0.4.3
nose==1.3.7
oauthlib==3.2.0
pathspec==0.9.0
platformdirs==2.4.1
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
soupsieve==2.3.1
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.9
zipp==3.8.0
